pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "resnet-cifar10"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "numpy",
    "pillow",
    "pyyaml",
    "tensorboard",
    "matplotlib",
    "seaborn",
    "thop",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.hatch.build.targets.wheel]
packages = ["src/resnet_cifar10"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
